telegram_bot/pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["cheapdramas <bileckijrostislav46@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.22.0"
dotenv = "^0.9.9"
pyjwt = "^2.10.1"
websockets = "^15.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
